pyproject.toml
[project]
name = "tongue-doctor"
version = "0.0.0"
description = "Multi-agent clinical reasoning research demonstration. Not a clinical tool."
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
  { name = "Rahman Shakoor", email = "rahman@yabagram.com" },
]
license = { text = "Proprietary" }
classifiers = [
  "Private :: Do Not Upload",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  # GCP
  "google-cloud-aiplatform>=1.71.0",
  "google-cloud-firestore>=2.19.0",
  "google-cloud-storage>=2.18.0",
  "google-cloud-secret-manager>=2.20.0",
  "google-cloud-logging>=3.11.0",
  "google-cloud-trace>=1.14.0",
  # LLM clients
  "anthropic[vertex]>=0.40.0",
  # Schemas + config
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.0",
  # HTTP + retries
  "httpx>=0.27.0",
  "tenacity>=9.0.0",
  # API server
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  # Retrieval
  "rank-bm25>=0.2.2",
  "cohere>=5.11.0",
  "tiktoken>=0.8.0",
  # Documents + images
  "pymupdf>=1.24.0",
  "pypdf>=5.0.0",
  "pillow>=11.0.0",
  "python-magic>=0.4.27",
  # Tooling
  "jinja2>=3.1.0",
  "structlog>=24.4.0",
  "typer>=0.13.0",
  # Observability
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-gcp-trace>=1.6.0",
]

[project.optional-dependencies]
benchmark = [
  "sentence-transformers>=3.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "hypothesis>=6.115.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
  "pre-commit>=4.0.0",
  "python-dotenv>=1.0.0",
  "freezegun>=1.5.0",
  "respx>=0.21.0",
  "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tongue_doctor"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "eval", "scripts"]

[tool.ruff.lint]
select = [
  "E", "F", "W",
  "I",
  "N",
  "UP",
  "B",
  "A",
  "C4",
  "RET",
  "SIM",
  "ARG",
  "RUF",
]
ignore = [
  "E501",
  "B008",
  "ARG002",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG", "PLR2004"]
"scripts/**" = ["T201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
files = ["src"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "google.cloud.*",
  "opentelemetry.exporter.cloud_trace",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=src/tongue_doctor",
  "--cov-report=term-missing",
  "--cov-fail-under=0",
]
markers = [
  "integration: integration tests requiring external services",
  "slow: slow tests",
]

[tool.coverage.run]
source = ["src/tongue_doctor"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
  "\\.\\.\\.",
]
